refactor: Move PaneHeader to its own folder
