+++ packages/seed-bible/seed-bible/components/PaneHeader/PaneHeader.css
@@ -1,0 +1,54 @@
+/* Chrome shown at the top of every pane: a title and a close button, and
+   nothing else. Floating panes use it as a drag handle to move the window;
+   side/fullscreen panes don't move, so it's just a static header there. */
+.sb-pane-header {
+  height: 42px;
+  flex: 0 0 42px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 8px;
+  padding: 0 12px;
+  border-bottom: 1px solid
+    color-mix(in srgb, var(--sb-font-color), transparent 35%);
+  cursor: default;
+  user-select: none;
+}
+
+.sb-pane-header-title {
+  flex: 1 1 auto;
+  min-width: 0;
+  font-weight: 700;
+  letter-spacing: 0.02em;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+/* Caller-provided header content (e.g. extension buttons), rendered between the
+   title and the close button. */
+.sb-pane-header-actions {
+  flex: 0 0 auto;
+  display: flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.sb-pane-header-close-button {
+  flex: 0 0 auto;
+  border: none;
+  background: transparent;
+  height: 2.5rem;
+  min-width: 2.5rem;
+  border-radius: 0.5rem;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  padding: 0;
+  color: inherit;
+}
+
+.sb-pane-header-close-button:hover {
+  background: color-mix(in srgb, var(--sb-secondary-color), transparent 65%);
+}
